requirements-minimal.txt
# Minimal requirements for Stock Price Prediction Tool
# Install with: pip install -r requirements-minimal.txt

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Collection
yfinance>=0.2.28

# Machine Learning  
scikit-learn>=1.3.0
xgboost>=2.0.0

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
joblib>=1.3.0
